requirements.txt
openpyxl==3.1.2
pandas==2.1.3
python-dotenv==1.0.0
python-telegram-bot==20.7
